ex) #이슈번호, #이슈번호

## 📝 요약(Summary)

변경 사항 및 관련 이슈에 대해 간단하게 작성해주세요. 어떻게보다 무엇을 왜 수정했는지 설명해주세요.

## 🛠️ PR 유형

어떤 변경 사항이 있나요?

- [ ] 새로운 기능 추가
- [ ] 버그 수정
- [ ] CSS 등 사용자 UI 디자인 변경
- [ ] 코드에 영향을 주지 않는 변경사항(오타 수정, 탭 사이즈 변경, 변수명 변경)
- [ ] 코드 리팩토링
- [ ] 주석

+++ Backend/IsCream/src/main/java/com/ssafy/iscream/calendar/service/CalendarService.java
@@ -8,10 +8,16 @@ import com.ssafy.iscream.calendar.dto.request.MemoUpdateReq;
 import com.ssafy.iscream.calendar.dto.response.CalendarGetDetailRes;
 import com.ssafy.iscream.calendar.dto.response.CalendarGetRes;
 import com.ssafy.iscream.calendar.repository.MemoRepository;
+import com.ssafy.iscream.children.domain.Child;
+import com.ssafy.iscream.common.exception.ErrorCode;
+import com.ssafy.iscream.common.exception.NotFoundException;
+import com.ssafy.iscream.common.exception.UnauthorizedException;
+import com.ssafy.iscream.common.response.ResponseData;
 import com.ssafy.iscream.htpTest.domain.HtpTest;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
 
+import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.Month;
 import java.util.ArrayList;
@@ -21,47 +27,72 @@ import java.util.Map;
 @Service
 @RequiredArgsConstructor
 public class CalendarService {
-    MemoRepository memoRepository;
-    // TODO: 예외처리 추가해야함
+    private final MemoRepository memoRepository;
 
-    public CalendarGetDetailRes getCalendarDetail(CalendarGetDetailReq calendarGetDetailReq) {
-        return null;
-    }
+    public void createMemo(Integer userId, MemoCreateReq memoCreateReq) {
 
-    public void createMemo(MemoCreateReq memoCreateReq) {
         Memo memo = Memo.builder()
+                .userId(userId)
                 .childId(memoCreateReq.getChildId())
                 .content(memoCreateReq.getMemo())
                 .build();
-
         memoRepository.save(memo);
     }
 
-    public void updateMemo(MemoUpdateReq memoUpdateReq) throws Exception{
+    public void updateMemo(Integer userId, MemoUpdateReq memoUpdateReq) {
+        // 존재 하지 않을 경우
         Memo memo = memoRepository.findById(memoUpdateReq.getMemoId())
-                .orElseThrow(() -> new Exception("메모를 찾을 수 없습니다."));
+                .orElseThrow(() -> new NotFoundException(
+                        new ResponseData<>(ErrorCode.DATA_NOT_FOUND.getCode(), ErrorCode.DATA_NOT_FOUND.getMessage(), null)));
+
+        // 권한 없는 경우
+        if (!memo.getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         memo.updateContent(memoUpdateReq.getMemo());
         memoRepository.save(memo);
     }
 
-    public void deleteMemo(int memoId) throws Exception{
+    public void deleteMemo(Integer userId, Integer memoId){
+        // 존재하지 않을 경우
         Memo memo = memoRepository.findById(memoId)
-                .orElseThrow(() -> new Exception("메모가 존재하지 않습니다."));
+                .orElseThrow(() -> new NotFoundException(
+                        new ResponseData<>(ErrorCode.DATA_NOT_FOUND.getCode(), ErrorCode.DATA_NOT_FOUND.getMessage(), null)));
+
+        // 권한 없는 경우
+        if (!memo.getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         memoRepository.deleteById(memoId);
     }
 
-    public List<Integer> getDaysByYearMonth(CalendarGetReq calendarGetReq) {
-        int year = calendarGetReq.getYearMonth().getYear();
-        Month month = calendarGetReq.getYearMonth().getMonth();
-        LocalDateTime startDate = LocalDateTime.of(year, month, 1, 0, 0);
+    public List<Integer> getDaysByYearMonth(Integer userId, CalendarGetReq calendarGetReq) {
+        LocalDateTime startDate = LocalDateTime.of(calendarGetReq.getYear(), calendarGetReq.getMonth(), 1, 0, 0);
         LocalDateTime endDate = startDate.plusMonths(1); // 다음 달 1일 (해당 월의 끝까지 포함)
 
-        List<Memo> memos = memoRepository.findByChildIdAndCreatedAtBetween( calendarGetReq.getChildId(), startDate, endDate);
+        List<Memo> memos = memoRepository.findByChildIdAndCreatedAtBetween(calendarGetReq.getChildId(), startDate, endDate);
         List<Integer> days = new ArrayList<>();
         for (Memo memo : memos) {
             days.add(memo.getCreatedAt().getDayOfMonth());
         }
+
+        // 권한 없는 경우
+        if (!memos.isEmpty() && !memos.get(0).getUserId().equals(userId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
         return days;
 
+    }
+
+    public Memo getByChildIdDate(Integer childId, LocalDate selectedDate) {
+        Memo memo = memoRepository.findByChildIdAndDate(childId, selectedDate).orElse(null);
+        if(memo != null && memo.getUserId().equals(childId)) {
+            throw new UnauthorizedException(new ResponseData<>(ErrorCode.DATA_FORBIDDEN_ACCESS.getCode(), ErrorCode.DATA_FORBIDDEN_ACCESS.getMessage(), null));
+        }
+
+        return memo;
     }
 }
